Remove `nextclade_datasets` directory and references
Based on Nextstrain meeting notes from 2023-09-28, we will be continuing
to use the centralized repo (https://github.com/nextstrain/nextclade_data)
for hosting the official Nextclade datasets since the Nextclade CLI
uses the repo for the central index.

This may change later, but with so many moving parts, this seems like
a better place to start.