ingest/curate.smk: Make the field map config more user friendly
When walking through the template with @kistlek, the field map config
param was a bit of a pain due to the spaces in the field names from
NCBI Datasets. I saw that @corneliusroemer made some improvements
in his nipah repo¹ to make this more user friendly so incorporating
those changes here.

¹ https://github.com/corneliusroemer/nipah/commit/16a0f94e3da9fc023a6a80bf0b396657c5fd428c

Co-authored-by: Cornelius Roemer <cornelius.roemer@gmail.com>