ingest: Add profile for Nextstrain automation
Adds a profile for Nextstrain automation that includes the config
and rules required to upload files to AWS S3.

I consider this a consolidation of all the different rules for
uploading to AWS S3 from the following pathogen workflows:
- https://github.com/nextstrain/ncov-ingest/blob/5aed3ed8b8e1bd0a8360f7135f1602df4e01d139/workflow/snakemake_rules/upload.smk
- https://github.com/nextstrain/monkeypox/blob/afb5513d9dea10e1993973a154844bf30cdbd335/ingest/workflow/snakemake_rules/upload.smk
- https://github.com/nextstrain/seasonal-flu/blob/06865c26a2d972252fde15735d001eaec261126f/profiles/upload/upload.smk

The files to upload are provided via the config, where the key is the
remote file and value is the local file. This allows us to upload a
local file to multiple remotes such as when we need to support multiple
compression formats.

The S3 URL and CloudFront domain are defined as top level configs to
allow for easy overrides via the Snakemake `--config` option.