ingest: simplify NCBI Datasets fields config
Instead of hard-coding the list of NCBI Datasets fields in the workflow,
just provide the list via the default config. This makes it easy to
customize which fields to include and makes it very obvious that
field_map config for the the curation pipeline is changing the names of
these NCBI fields.

This includes a change in the `format_ncbi_dataset_report` rule to use
the provided fields as the header so that we do not have to do a
separate renaming of the NCBI column names back to the computer friendly
mneumonics.