Rename "profiles" to "build-configs"
To avoid confusion with Snakemake's Profiles¹, rename the "profiles"
directory to "build-configs" based on discussion in
https://github.com/nextstrain/pathogen-repo-template/pull/27.

This is expected to include subdirectories for specific build configs
that include both config params and custom Snakemake rules that extend
and/or override the core workflow.

This also allows us to rename the build specific configs from
"defaults.yaml" to "config.yaml" since they no longer clash with
Snakemake Profile configuration files.

¹ https://snakemake.readthedocs.io/en/stable/executing/cli.html#profiles