ingest: Merge Nextclade metadata with `augur merge`
This construction reads a bit clearer and cleaner.  It's also a good
example of how to use `augur merge`.

The limitation on non-seekable streams means the workflow now uses
additional transient disk space, but it typically shouldn't be an issue.
The way Augur's slow start up time impacts `augur merge` also
contributes to a longer rule execution time, but it should be negligible
in the context of the larger workflow and presumably we'll fix the slow
start up eventually.¹

The output is semantically identical but has some syntactic changes re:
quoting.  It's worth noting that the pre-existing TSV format was _not_
IANA TSV, despite it (still) being treated as such in a few places, but
was (and remains) a CSV-like TSV with some quoted fields.  We really
need to sort out our TSV formats³, but that's for a larger project.

¹ <https://github.com/nextstrain/augur/issues/1628>
² <https://github.com/nextstrain/augur/pull/1565>
³ <https://github.com/nextstrain/augur/issues/1566>

Ported-from: <https://github.com/nextstrain/measles/commit/4d73b7fd7acc4428e599908b7bb2450086797b27>
Related-to: <https://github.com/nextstrain/measles/pull/52>
Related-to: <https://github.com/nextstrain/pathogen-repo-guide/issues/65>