merge_inputs: generalize input_* functions for wildcards
Use `rules` variable for `input_*` functions to ensure that the returned merged 
path stay in-sync with the `merge_*` rules outputs. This makes it easier to 
adapt the template to use wildcards.

Technically `_gather_inputs`, `input_metadata`, and `input_sequences` are 
generalized enough to vendor as shared functions. However, I think it's easier 
to understand how the inputs are handled when they are colocated with the 
`merge_*` rules. If we use Snakemake modules to use the `merge_*` rules, then 
this whole file can be vendored, but that's a bigger change with unknown 
pitfalls so I'll punt on it for now.