Add support for `nextstrain run`
Workflow changes to support running the workflows with the 
`nextstrain run` command.

I specifically chose _not_ to declare compatibility for `nextstrain run` in 
the `nextstrain-pathogen.yaml` because the phylogenetic and nextclade workflows 
are incomplete in this guide. Once workflow specific compatibility is 
supported,¹ we can declare compatibility for the ingest workflow.

¹ <https://github.com/nextstrain/cli/pull/472>